finish commit

+++ ReclutamientoBackend/Dto/CandidatoExperiencia/CandidatoExperienciaCreateDto.cs
@@ -6,7 +6,7 @@
         public string Company { get; set; }
         public string Job { get; set; }
         public string Description { get; set; }
-        public float Salary { get; set; }
+        public Decimal Salary { get; set; }
         public DateTime BeginDate { get; set; }
         public DateTime? EndDate { get; set; }
         public DateTime InsertDate { get; set; }
